+++ lib/redmine_issue_view_columns/project_helper_patch.rb
@@ -1,5 +1,7 @@
 module RedmineIssueViewColumns
   module ProjectHelperPatch
+    include IssueViewColumnsHelper
+
     def project_settings_tabs
       super.tap do |tabs|
         tabs << {
@@ -9,6 +11,14 @@ module RedmineIssueViewColumns
           label: :issue_view_columns_settings,
         } if User.current.allowed_to?(:manage_issue_view_columns, @project) &&
              @project.module_enabled?(:issue_view_columns)
+
+        tabs << {
+          name: "issue_view_columns_relations",
+          action: :issue_view_columns,
+          partial: "issue_view_columns/relation_types",
+          label: :label_issue_view_columns_relation_types_tab,
+        } if (User.current.allowed_to?(:manage_issue_view_columns, @project) || User.current.admin?) &&
+             additional_relation_type_keys.any?
       end
     end
   end